assistant-gazelle-web/requirements.txt
# Assistant Gazelle Web - Dependencies
# Installation: pip install -r requirements.txt

# Core Framework
Flask==3.0.0
flask-cors==4.0.0

# Base de données - Supabase (PostgreSQL) - PRIORITÉ
psycopg2-binary==2.9.9

# Base de données - SQLite (développement local)
# Note: sqlite3 est inclus dans Python standard, pas besoin d'installer

# Base de données - SQL Server (fallback Windows uniquement)
pyodbc==5.0.1

# Configuration (variables d'environnement)
python-dotenv==1.0.0

# OpenAI Integration (si nécessaire)
openai==1.6.1

# Authentication & Security
PyJWT==2.8.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Production Server
gunicorn==21.2.0  # Pour Linux/Mac/Render
waitress==2.1.2   # Pour Windows (optionnel)

# HTTP Requests (pour API Gazelle)
requests==2.31.0

